[all] rename SimpleKiwi to Kiwi
