+++ Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/Aboveground/SimpleAsymmetricZOI/ReferenceFiles/SimpleAsymmetricZOI/Population_t_0002000000.0.csv
