+++ Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/Aboveground/SimpleAsymmetricZOI/SimpleAsymmetricZOI.xml
@@ -2,16 +2,23 @@
     <random_seed>643879</random_seed>
     <resources>
         <aboveground>
-            <type> Default </type>
+            <type> SimpleAsymmetricZOI </type>
+            <domain>
+                <x_1> 0 </x_1>
+                <y_1> 0 </y_1>
+                <x_2> 22 </x_2>
+                <y_2> 22 </y_2>
+            <x_resolution> 88 </x_resolution>
+            <y_resolution> 88 </y_resolution>
+            </domain>
         </aboveground>
         <belowground>
             <type> Default </type>
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> SimpleKiwi </type>
-        <mortality>Random</mortality>
-        <probability>0.0016 </probability>
+        <type> Kiwi </type>
+        <mortality>NoGrowth</mortality>
     </plant_dynamics>
     <population>
         <group>
@@ -20,29 +27,30 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/SimpleKiwi/initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/initial_population.csv </filename>
             </distribution>
         </group>
     </population>
     <time_loop>
         <type> Simple </type>
         <t_start> 0 </t_start>
-        <t_end> 2e6 </t_end>
+        <t_end> 5e8 </t_end>
         <delta_t> 1e6 </delta_t>
     </time_loop>
     <visualization>
         <type> NONE </type>
     </visualization>
     <output>
-        <type> OneTimestepOneFile </type>
+         <type> OneTimestepOneFile </type>
         <output_each_nth_timestep>1</output_each_nth_timestep>
-        <output_times> [2e6] </output_times>
+        <output_times> [2e6, 5e8] </output_times>
         <allow_previous_output>True</allow_previous_output>
         <output_dir>Benchmarks/TestOutputs/</output_dir>
         <geometry_output> r_stem </geometry_output>
         <geometry_output> height </geometry_output>
         <growth_output> growth </growth_output>
-        <growth_output> ag_factor </growth_output>
-        <growth_output> bg_factor </growth_output>
+        <growth_output> ag_resources </growth_output>
+        <growth_output> bg_resources </growth_output>
     </output>
 </MangaProject>
+
