+++ Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/Belowground/FixedSalinity/FixedSalinity_CI.xml
@@ -5,11 +5,14 @@
             <type> Default </type>
         </aboveground>
         <belowground>
-            <type> Default </type>
+            <type> FixedSalinity </type>
+            <min_x>0</min_x>
+            <max_x>22</max_x>
+            <salinity>0.025 0.035</salinity>
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> SimpleKiwi </type>
+        <type> Kiwi </type>
         <mortality>NoGrowth</mortality>
     </plant_dynamics>
     <population>
@@ -19,23 +22,23 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/SimpleKiwi/initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/initial_population.csv </filename>
             </distribution>
         </group>
     </population>
     <time_loop>
         <type> Simple </type>
         <t_start> 0 </t_start>
-        <t_end> 5e8 </t_end>
+        <t_end> 2e6 </t_end>
         <delta_t> 1e6 </delta_t>
     </time_loop>
     <visualization>
         <type> NONE </type>
     </visualization>
     <output>
-        <type> OneTimestepOneFile </type>
+         <type> OneTimestepOneFile </type>
         <output_each_nth_timestep>1</output_each_nth_timestep>
-        <output_times> [2e6, 5e8] </output_times>
+        <output_times> [2e6] </output_times>
         <allow_previous_output>True</allow_previous_output>
         <output_dir>Benchmarks/TestOutputs/</output_dir>
         <geometry_output> r_stem </geometry_output>
