+++ Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/Default/Default_CI.xml
@@ -9,9 +9,8 @@
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> SimpleKiwi </type>
-        <mortality>RandomGrowth</mortality>
-        <k_die> 1e-9 </k_die>
+        <type> Kiwi </type>
+        <mortality>NoGrowth</mortality>
     </plant_dynamics>
     <population>
         <group>
@@ -20,7 +19,7 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/SimpleKiwi/initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/initial_population.csv </filename>
             </distribution>
         </group>
     </population>
@@ -34,7 +33,7 @@
         <type> NONE </type>
     </visualization>
     <output>
-        <type> OneTimestepOneFile </type>
+         <type> OneTimestepOneFile </type>
         <output_each_nth_timestep>1</output_each_nth_timestep>
         <output_times> [2e6] </output_times>
         <allow_previous_output>True</allow_previous_output>
@@ -46,3 +45,4 @@
         <growth_output> bg_factor </growth_output>
     </output>
 </MangaProject>
+
