+++ Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/Mortality/Random/Random.xml
@@ -9,10 +9,9 @@
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> SimpleKiwi </type>
-        <mortality>Memory</mortality>
-        <threshold> 0.005 </threshold>
-        <period> 31557600 </period>
+        <type> Kiwi </type>
+        <mortality>Random</mortality>
+        <probability>0.016 </probability>
     </plant_dynamics>
     <population>
         <group>
@@ -21,14 +20,14 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/SimpleKiwi/initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/initial_population.csv </filename>
             </distribution>
         </group>
     </population>
     <time_loop>
         <type> Simple </type>
         <t_start> 0 </t_start>
-        <t_end> 2e6 </t_end>
+        <t_end> 5e8 </t_end>
         <delta_t> 1e6 </delta_t>
     </time_loop>
     <visualization>
@@ -37,7 +36,7 @@
     <output>
         <type> OneTimestepOneFile </type>
         <output_each_nth_timestep>1</output_each_nth_timestep>
-        <output_times> [2e6] </output_times>
+        <output_times> [2e6, 5e8] </output_times>
         <allow_previous_output>True</allow_previous_output>
         <output_dir>Benchmarks/TestOutputs/</output_dir>
         <geometry_output> r_stem </geometry_output>
