+++ Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/Mortality/RandomGrowth/RandomGrowth.xml
@@ -9,9 +9,9 @@
         </belowground>
     </resources>
     <plant_dynamics>
-        <type> SimpleKiwi </type>
-        <mortality>Random</mortality>
-        <probability>0.016 </probability>
+        <type> Kiwi </type>
+        <mortality>RandomGrowth</mortality>
+        <k_die> 1e-9 </k_die>
     </plant_dynamics>
     <population>
         <group>
@@ -20,7 +20,7 @@
             <distribution>
                 <type> GroupFromFile </type>
                 <n_recruitment_per_step> 0 </n_recruitment_per_step>
-                <filename> Benchmarks/ModuleBenchmarks/PlantModules/SimpleKiwi/initial_population.csv </filename>
+                <filename> Benchmarks/ModuleBenchmarks/PlantModules/Kiwi/initial_population.csv </filename>
             </distribution>
         </group>
     </population>
