+++ PlantModelLib/Kiwi/__init__.py
@@ -6,4 +6,4 @@ Created on Thu Nov  8 15:25:03 2018
 @author: bathmann
 """
 
-from .SimpleKiwi import SimpleKiwi
+from .Kiwi import Kiwi
